// tools/types/custom-block.ts
// Generated by https://quicktype.io

export interface CustomBlock {
  format_version: string;
  'minecraft:block': MinecraftBlock;
}

export interface MinecraftBlock {
  description: Description;
  components: Components;
}

export interface Components {
  'minecraft:creative_category': MinecraftCreativeCategory;
  'minecraft:destroy_time': number;
  'minecraft:explosion_resistance': number;
  'minecraft:friction': number;
  'minecraft:flammable': MinecraftFlammable;
  'minecraft:map_color': string;
  'minecraft:block_light_absorption': number;
  'minecraft:block_light_emission': number;
}

export interface MinecraftCreativeCategory {
  category: string;
}

export interface MinecraftFlammable {
  flame_odds: number;
  burn_odds: number;
}

export interface Description {
  identifier: string;
  is_experimental: boolean;
  register_to_creative_menu: boolean;
}