// tools/types/manifest.ts
// Generated by https://quicktype.io

export interface Manifest {
  format_version: number;
  header: Header;
  modules: Module[];
  dependencies: Dependency[];
  metadata: Metadata;
}

export interface Dependency {
  uuid: string;
  version: number[];
}

export interface Header {
  description: string;
  name: string;
  uuid: string;
  version: number[];
  min_engine_version: number[];
}

export interface Metadata {
  authors: string[];
  license: string;
  url?: string;
}

export interface Module {
  description: string;
  type: string;
  uuid: string;
  version: number[];
}